--- src/main/java/ClassicalMusic.java
@@ -1,20 +1,0 @@
-public class ClassicalMusic implements Music {
-    private ClassicalMusic() {}
-
-    public static ClassicalMusic getClassicalMusic() {
-        return new ClassicalMusic();
-    }
-
-    @Override
-    public String getSong() {
-        return "Hungarian Rhapsody";
-    }
-    public void doMyInit() {
-        System.out.println("Doing my initialization");
-    }
-
-    public void doMyDestroy() {
-        System.out.println("Doing my destruction");
-    }
-
-}
