+++ src/main/java/Task21/ClassicalMusic.java
@@ -1,0 +1,11 @@
+package Task21;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class ClassicalMusic implements Music {
+    @Override
+    public String getSong() {
+        return "Hungarian Rhapsody";
+    }
+}
