+++ src/main/java/Task21/Music.java
@@ -1,3 +1,5 @@
+package Task21;
+
 public interface Music {
     String getSong();
 }
