+++ src/main/java/Task21/MusicPlayer.java
@@ -1,0 +1,15 @@
+package Task21;
+
+import Task21.Music;
+
+public class MusicPlayer {
+    private Music music;
+
+    public MusicPlayer(Music music) {
+        this.music = music;
+    }
+
+    public void playMusic() {
+        System.out.println("Playing: " + music.getSong());
+    }
+}
