+++ src/main/java/Task21/RockMusic.java
@@ -1,3 +1,9 @@
+package Task21;
+
+import Task21.Music;
+import org.springframework.stereotype.Component;
+
+@Component
 public class RockMusic implements Music {
     @Override
     public String getSong() {
