+++ src/main/java/Task21/TestSpring.java
@@ -1,0 +1,27 @@
+package Task21;
+
+import Task21.Music;
+import Task21.MusicPlayer;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+
+public class TestSpring {
+    public static void main(String[] args) {
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(
+                "applicationContext.xml"
+        );
+
+        Music music = context.getBean("rockMusic", Music.class);
+
+        MusicPlayer musicPlayer = new MusicPlayer(music);
+
+        musicPlayer.playMusic();
+
+        Music music2 = context.getBean("classicalMusic", Music.class);
+
+        MusicPlayer classicalMusicPlayer = new MusicPlayer(music2);
+
+        classicalMusicPlayer.playMusic();
+
+        context.close();
+    }
+}
