--- src/main/java/TestSpring.java
@@ -1,18 +1,0 @@
-import org.springframework.context.support.ClassPathXmlApplicationContext;
-
-public class TestSpring {
-    public static void main(String[] args) {
-        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(
-                "applicationContext.xml"
-        );
-
-        MusicPlayer musicPlayer = context.getBean("MusicPlayer", MusicPlayer.class);
-        musicPlayer.playMusic();
-
-        System.out.println(musicPlayer.getName());
-        System.out.println(musicPlayer.getVolume()
-        );
-
-        context.close();
-    }
-}
